# docs/openapi.yaml
openapi: 3.1.0
info:
  title: devin-task-board API
  version: 1.0.0
  description: タスク管理アプリケーション（Jira ミニ版）の REST API 仕様
servers:
  - url: http://localhost:3000
    description: ローカル開発環境
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth.js v5 が発行する JWT セッショントークン
  schemas:
...
    ErrorCode:
      type: string
      description: |
        アプリケーション全体で使用されるエラーコード一覧:
        - VALIDATION_ERROR: リクエストのバリデーション失敗（400）
        - UNAUTHORIZED: 未認証（401）
        - FORBIDDEN: 権限不足（403）
        - NOT_FOUND: リソース未発見（404）
        - CONFLICT: リソース競合（409）
        - INTERNAL_ERROR: サーバー内部エラー（500）
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: 人間が読めるエラーメッセージ
              examples:
                - タイトルは必須です
    Pagination:
      type: object
      required:
        - page
        - perPage
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
          description: 現在のページ番号
          examples:
            - 1
        perPage:
          type: integer
          minimum: 1
          maximum: 100
          description: 1ページあたりの件数
          examples:
            - 20
        total:
          type: integer
          minimum: 0
          description: 全件数
          examples:
            - 100
        totalPages:
          type: integer
          minimum: 0
          description: 全ページ数
          examples:
            - 5

    # ── User ──
...
    UserSummary:
      type: object
      description: ユーザーの要約情報（リレーション展開用）
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: ユーザーID（cuid）
        name:
          type: string
          description: 表示名
        avatarUrl:
          type:
            - string
            - "null"
          description: プロフィール画像URL

    # ── Project ──
...
    Task:
      type: object
      required:
        - id
        - taskNumber
        - title
        - status
        - priority
        - projectId
        - reporterId
        - sortOrder
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: タスクID（cuid）
        taskNumber:
          type: integer
          minimum: 1
          description: プロジェクト内連番
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: タイトル
        description:
          type:
            - string
            - "null"
          description: Markdown 対応の説明
        status:
          type: string
          enum:
            - BACKLOG
            - TODO
            - IN_PROGRESS
            - IN_REVIEW
            - DONE
          description: ステータス
        priority:
          type: string
          enum:
            - URGENT
            - HIGH
            - MEDIUM
            - LOW
            - NONE
          description: 優先度
        projectId:
          type: string
        assigneeId:
          type:
            - string
            - "null"
          description: 担当者のユーザーID
        reporterId:
          type: string
          description: 起票者のユーザーID
        parentTaskId:
          type:
            - string
            - "null"
          description: 親タスクID（サブタスク用）
        dueDate:
          type:
            - string
            - "null"
          format: date-time
          description: 期限
        startDate:
          type:
            - string
            - "null"
          format: date-time
          description: 開始日（ガントチャート用）
        estimatedHours:
          type:
            - number
            - "null"
          minimum: 0
          description: 見積もり工数
        actualHours:
          type:
            - number
            - "null"
          minimum: 0
          description: 実績工数
        sortOrder:
          type: integer
          description: カンバン内並び順
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskWithRelations:
      description: リレーション情報を含むタスク（一覧・詳細取得時のレスポンス）
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          required:
            - reporter
            - taskCategories
          properties:
            assignee:
              oneOf:
                - $ref: "#/components/schemas/UserSummary"
                - type: "null"
              description: 担当者情報
            reporter:
              $ref: "#/components/schemas/UserSummary"
              description: 起票者情報
            taskCategories:
              type: array
              items:
                type: object
                required:
                  - category
                properties:
                  category:
                    $ref: "#/components/schemas/Category"
              description: 紐づくカテゴリ一覧
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: タイトル（必須）
        description:
          type: string
          description: 説明（Markdown）
        status:
          type: string
          enum:
            - BACKLOG
            - TODO
            - IN_PROGRESS
            - IN_REVIEW
            - DONE
          default: BACKLOG
        priority:
          type: string
          enum:
            - URGENT
            - HIGH
            - MEDIUM
            - LOW
            - NONE
          default: NONE
        assigneeId:
          type: string
          description: 担当者のユーザーID
        dueDate:
          type: string
          format: date-time
          description: 期限
        startDate:
          type: string
          format: date-time
          description: 開始日
        estimatedHours:
          type: number
          description: 見積もり工数
        categoryIds:
          type: array
          items:
            type: string
          description: カテゴリIDの配列
...
    Category:
      type: object
      required:
        - id
        - name
        - color
        - projectId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: カテゴリID（cuid）
        name:
          type: string
          description: カテゴリ名
        color:
          type: string
          description: OKLCH カラー値
        projectId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: プロジェクトID
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: ページ番号
    PerPageParam:
      name: perPage
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 1ページあたりの件数（1〜100）
  responses:
    BadRequest:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: VALIDATION_ERROR
              message: タイトルは必須です
    Unauthorized:
      description: 未認証
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: UNAUTHORIZED
              message: 認証が必要です
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: FORBIDDEN
              message: この操作を行う権限がありません
    NotFound:
      description: リソース未発見
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: NOT_FOUND
              message: リソースが見つかりません
...
    InternalServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: INTERNAL_ERROR
              message: サーバーエラーが発生しました
paths:
...
  /api/projects/{id}/tasks:
    get:
      tags:
        - Tasks
      summary: タスク一覧
      description: プロジェクト内のタスク一覧を取得する。ページネーション・フィルター・ソート対応
      operationId: listTasks
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - name: status
          in: query
          schema:
            type: string
            enum:
              - BACKLOG
              - TODO
              - IN_PROGRESS
              - IN_REVIEW
              - DONE
          description: ステータスでフィルター
        - name: priority
          in: query
          schema:
            type: string
            enum:
              - URGENT
              - HIGH
              - MEDIUM
              - LOW
              - NONE
          description: 優先度でフィルター
        - name: assigneeId
          in: query
          schema:
            type: string
          description: 担当者でフィルター
        - name: categoryId
          in: query
          schema:
            type: string
          description: カテゴリでフィルター
        - name: sortBy
          in: query
          schema:
            type: string
            enum:
              - taskNumber
              - title
              - status
              - priority
              - dueDate
              - createdAt
            default: createdAt
          description: ソートカラム
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: ソート順
      responses:
        "200":
          description: タスク一覧取得成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskWithRelations"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                    description: "※ 現在の実装では未返却。ページネーション実装後に required へ昇格予定"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Tasks
      summary: タスク作成
      description: プロジェクト内にタスクを作成する。タスク番号が自動採番される
      operationId: createTask
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: タスク作成成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/TaskWithRelations"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...